Make indexing strategy implementations private
