+++ include/ccf/indexing/strategies/visit_each_entry_in_map.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/byte_vector.h"
 #include "ccf/indexing/strategy.h"
-#include "kv/store.h"
 
 namespace ccf::indexing::strategies
 {
@@ -18,40 +18,20 @@ namespace ccf::indexing::strategies
 
     virtual void visit_entry(
       const ccf::TxID& tx_id,
-      const kv::serialisers::SerialisedEntry& k,
-      const kv::serialisers::SerialisedEntry& v) = 0;
+      const ccf::ByteVector& k,
+      const ccf::ByteVector& v) = 0;
 
   public:
     VisitEachEntryInMap(
       const std::string& map_name_,
-      const std::string& strategy_prefix = "VisitEachEntryIn") :
-      Strategy(fmt::format("{} {}", strategy_prefix, map_name_)),
-      map_name(map_name_)
-    {}
+      const std::string& strategy_prefix = "VisitEachEntryIn");
+
+    virtual ~VisitEachEntryInMap() = default;
 
     void handle_committed_transaction(
-      const ccf::TxID& tx_id, const kv::StorePtr& store) override
-    {
-      // NB: Get an untyped view over the map with the same name. This saves
-      // deserialisation here, where we hand on the raw key and value.
-      auto tx = store->create_read_only_tx();
-      auto handle = tx.ro<kv::untyped::Map>(map_name);
+      const ccf::TxID& tx_id, const kv::StorePtr& store) override;
+    std::optional<ccf::SeqNo> next_requested() override;
 
-      handle->foreach([this, &tx_id](const auto& k, const auto& v) {
-        visit_entry(tx_id, k, v);
-        return true;
-      });
-      current_txid = tx_id;
-    }
-
-    std::optional<ccf::SeqNo> next_requested() override
-    {
-      return current_txid.seqno + 1;
-    }
-
-    ccf::TxID get_indexed_watermark() const
-    {
-      return current_txid;
-    }
+    ccf::TxID get_indexed_watermark() const;
   };
 }
