+++ src/ds/map_serializers.h
@@ -2,6 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/byte_vector.h"
 #include "ccf/ccf_assert.h"
 #include "serialized.h"
 
@@ -30,7 +31,7 @@ namespace map
 
   namespace serialisers
   {
-    using SerialisedEntry = llvm_vecsmall::SmallVector<uint8_t, 8>;
+    using SerialisedEntry = ccf::ByteVector;
   }
 
   namespace untyped
