+++ src/ds/test/hash_bench.cpp
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 
+#include "ccf/byte_vector.h"
 #include "ds/hash.h"
 
 #define PICOBENCH_IMPLEMENT_WITH_MAIN
