+++ src/kv/serialised_entry.h
@@ -1,13 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #pragma once
+#include "ccf/byte_vector.h"
 
-#include <nlohmann/json.hpp>
-#include <small_vector/SmallVector.h>
+#define FMT_HEADER_ONLY
+#include <fmt/format.h>
 
 namespace kv::serialisers
 {
-  using SerialisedEntry = llvm_vecsmall::SmallVector<uint8_t, 8>;
+  using SerialisedEntry = ccf::ByteVector;
 }
 
 namespace fmt
